% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get an ACAPS dataset}
\usage{
get_dataset(endpoint_url, token = .token(), ...)
}
\arguments{
\item{endpoint_url}{The API endpoint for the data you want to retrieve.}

\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
returns the requested data as an R dataframe.
}
\description{
A general high level function to retrieve the data provided by the ACAPS API
endpoint. Additional parameters can be specified with the request. The resulting
R dataframe is 'flattened', meaning that fields containing multiple values are converted
to a single string.
}
