% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_access.R
\name{get_dataset_access}
\alias{get_dataset_access}
\title{Get ACAPS Humanitarian Access Dataset}
\usage{
get_dataset_access(
  endpoint = c("{date}", "log", "events", "events-expanded-admin", "weighted-scores",
    "infogap-weighted-scores"),
  date = NULL,
  token = .token(),
  ...
)
}
\arguments{
\item{endpoint}{Specific endpoint to access. The most common use case is to
access the {date} endpoint. The API expects {date} to be in a specific format, for January
2023 this would look like: Jan2023. The earliest possible date is Jun2020.}

\item{date}{An additional date argument is required when accessing the 'weighted-scores'
or 'infogap-weighted-scores' endpoint. For other endpoints this can be left blank.}

\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
A specialized function to access the ACAPS Humanitarian Access group of endpoints.
}
