% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_countries.R
\name{get_dataset_countries}
\alias{get_dataset_countries}
\title{Get ACAPS Countries Dataset}
\usage{
get_dataset_countries(token = .token(), ...)
}
\arguments{
\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
A specialized function to call the ACAPS Countries API endpoint.
}
