% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_crises.R
\name{get_dataset_crises}
\alias{get_dataset_crises}
\title{Get ACAPS Crises Dataset}
\usage{
get_dataset_crises(endpoint = NULL, token = .token(), ...)
}
\arguments{
\item{endpoint}{The API endpoint to request. One specification is available: 'crisis-activation-log'.
If left blank, the function will call the general ACAPS crises endpoint.}

\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
Specialized R function for calls to the ACAPS Crises group of endpoints.
}
