% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_severity.R
\name{get_dataset_severity}
\alias{get_dataset_severity}
\title{Get ACAPS INFORM Severity Index Dataset}
\usage{
get_dataset_severity(
  endpoint = c("{date}", "country-log", "log", "impact-of-crisis",
    "conditions-of-people-affected", "complexity", "reliability", "core-indicators",
    "country-indicators"),
  date = NULL,
  token = .token(),
  ...
)
}
\arguments{
\item{endpoint}{Specific endpoint to access. The most common use case is to
access the {date} endpoint. The API expects {date} to be in a specific format, for January
2023 this would look like: Jan2023. The earliest possible date is Jun2020.}

\item{date}{An additional date argument is required when accessing the 'weighted-scores'
or 'infogap-weighted-scores' endpoint. For other endpoints this can be left blank.}

\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
A specialized function to call the ACAPS INFORM Severity Index group of endpoints.
}
