% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_single_endpoint.R
\name{get_dataset_single_endpoint}
\alias{get_dataset_single_endpoint}
\alias{get_dataset_infolandscape}
\alias{get_dataset_countries}
\alias{get_dataset_risklist}
\alias{get_dataset_govmeasures}
\alias{get_dataset_secondaryimpacts}
\alias{get_dataset_dailymonitoring}
\alias{get_dataset_protindicators}
\title{Get ACAPS dataset from a single endpoint}
\usage{
get_dataset_infolandscape(token = .token(), ...)

get_dataset_countries(token = .token(), ...)

get_dataset_risklist(token = .token(), ...)

get_dataset_govmeasures(token = .token(), ...)

get_dataset_secondaryimpacts(token = .token(), ...)

get_dataset_dailymonitoring(token = .token(), ...)

get_dataset_protindicators(token = .token(), ...)
}
\arguments{
\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
These specialized functions can be used to call ACAPS datasets from categories
with only a single API endpoint. These functions do not require any arguments besides
your API access token. They accept optional additional parameters
}
