% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_ukraine.R
\name{get_dataset_ukraine}
\alias{get_dataset_ukraine}
\title{Get ACAPS Ukraine dataset}
\usage{
get_dataset_ukraine(
  endpoint = c("subnational-access", "access-events", "damages", "master-dataset",
    "regional-dataset"),
  date = NULL,
  token = .token(),
  ...
)
}
\arguments{
\item{endpoint}{Specific endpoint to access.}

\item{date}{An additional date argument is required when accessing the 'subnational-access',
'master-dataset' or  'regional-dataset' endpoint. The API expects {date} to
be in a specific format, for January 2023 this would look like: Jan2023.
The earliest possible date is Jun2020. For other endpoints this can be left blank.}

\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
A specialized function to call the ACAPS Ukraine group of endpoints.
}
