% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_yemen.R
\name{get_dataset_yemen}
\alias{get_dataset_yemen}
\title{Get ACAPS Yemen monitoring dataset}
\usage{
get_dataset_yemen(endpoint = NULL, token = .token(), ...)
}
\arguments{
\item{endpoint}{Specific endpoint to access. There is one (optional) endpoint: 'indicators'}

\item{token}{Your authentication token. If left empty, the environment variable
created with \code{set_env_token()} will be used.}

\item{...}{Additional parameters to be passed along with the API request. Parameters
must be provided as name-value pairs.}
}
\value{
The requested dataset is returned as an R dataframe.
}
\description{
A specialized function to call the ACAPS Yemen Monitoring group of endpoints.
}
