% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get an API token for your account}
\usage{
get_token(username, password)
}
\arguments{
\item{username}{Your ACAPS account username (usually an e-mail address).}

\item{password}{Your ACAPS account password.}
}
\value{
This returns your API authentication token as character string.
}
\description{
This function is used to retrieve an API authentication token for your account
from the ACAPS API. A valid authentication token (and by extension an account)
is required in order to use the ACAPS API.
}
