% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_response}
\alias{parse_response}
\title{Parses the content of an API response}
\usage{
parse_response(response)
}
\arguments{
\item{response}{The response object returned by the API.}
}
\value{
The function returns the JSON response parsed as an R object
}
\description{
Parses the content of an API response
}
