% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env_token.R
\name{set_env_token}
\alias{set_env_token}
\title{Set ACAPS authentication token as env var}
\usage{
set_env_token(token, silent_overwrite = FALSE)
}
\arguments{
\item{token}{Your ACAPS authentication token}

\item{silent_overwrite}{If set to TRUE, function will always overwrite existing value. If FALSE,
will prompt the user if they want to overwrite the existing value. Set to FALSE by default.}
}
\description{
This function is used to store your ACAPS authentication token as en environment
variable. This allows the package functions to access your authentication token
without putting the token value itself in scripts. All functions that require a
token will look for this environment variable by default.
}
